requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
discord.py[voice]>=2.2.3
emoji==2.0.0
emoji-country-flag==1.3.1
idna==3.3
langdetect==1.0.9
mojimoji==0.0.12
multidict==6.0.2
pycparser==2.21
PyNaCl==1.5.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2
typing_extensions==4.1.1
urllib3==1.26.9
wanakana==1.0.1
yarl==1.7.2
